+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Max collisions into AbstractInteractor
+
 ## [5.1.0] - 2026-08-03
 ### Changed
 - Unseal AreaTrigger and AreaTriggerUnityEvent